split test for bundleMapToList and bundleListToMap
